requirements.txt
fastapi
requests
python-dotenv
google-generativeai
